Make legacy sidebar code manageable
* I first wanted to rewrite to hooks in one go, since this is the most
confusing class component we have. However it's too complex to do at
once. This commit does a few low hanging fruit improvements that are
easy to review without needing to go too much into the more confusing
parts. The most confusing being the 2 tightly coupled HOCs WithPostMeta
and FromThemeOptions.
* Move some of the logic to separate components so that it's not all
inside of CampaignSidebar.
* Return early a few times.
* Some naming improvements.
* Add documentation for some obscure parts.
